requirements.txt
groq
faiss-cpu
sentence-transformers
langtrace-python-sdk
opentelemetry-exporter-otlp